REFACTOR: MyselfTaughtCurriculum components
	child components now has their own folders and files.
