+++ components/CustomWindows/MySelfTaughtCurriculum/SectionArea/sectionsArea.tsx
@@ -1,0 +1,11 @@
+import { ReactNode } from "react"
+
+export default function SectionsArea(props: {
+    children: ReactNode
+}) {
+    return (
+        <ul>
+            {props.children}
+        </ul>
+    )
+}
